Finalize settings page
